[native] implement onPressSetAvatar function
Summary:
Implement the `onPressSetAvatar` function. This diff handles the user experience for when a user presses the save avatar button.

Here's an outline of the experience from a high level:

  # We attempt to call the update user avatar endpoint
  # While we are making this attempt, both the save and reset button will be disabled and a spinner will render on top of the emoji avatar
  # If we are successful we will show a toast message saying the avatar has been updated
  # If we fail then an alert will display telling the user something went wrong and to try again later

Depends on D7403

Test Plan:
Please watch the demo video to see the changes I made

Success case:
{F469282}

Failed case:
{F469291}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7359