[keyserver] Get rid of `versions` table and related endpoints
Summary:
We're not using this table and don't have any immediate plans to do so. If we do need to get the data `versions` was designed to store, we can easily get itfrom GitHub/AppStoreConnect/PlayStore APIs.

Context:
https://linear.app/comm/issue/ENG-3633/compute-equivalent-of-versions-table-using-ghappstoreplaystore-apis

https://linear.app/comm/issue/ENG-2428/compute-equivalent-of-keyserver-versions-table-with

Test Plan:
Ran migration and my `versions` table is gone:

Reviewers:
ashoat, ginsu