[native] Block displaying edit button when in edit mode
Summary: We want to prevent a user from losing edited changes by removing the edit button in the tooltip when the user is in edit mode.

Test Plan: Turned on the app on iOS. Entered edit mode. Checked if the edit button doesn't show anymore. Exited edit mode. Checked if the button appeared again.

Reviewers: michal, inka, ginsu