[keyserver] Migration to reset all PRIVATE thread names
Summary:
For all `PRIVATE` threads where the thread name is the member's username, this will clear out the thread name.

Once D7453 is widely deployed to clients, this will have the effect of allowing ENS resolution for PRIVATE thread names. For older clients without D7453, this would result in the threads being titled "just you".

To reduce the likelihood that a user encounters the "just you" case, I'll wait until we have a build with D7453 live in prod for a bit before landing this diff.

Depends on D7454

Test Plan: Tested the migration in my local environment. Confirmed it unsets the thread names correctly

Reviewers: tomek, atul

Differential Revision: https://phab.comm.dev/D7455