[native] fix Android cancel action sheet crash
Summary: While testing the edit avatar flow for android, I encountered this bug where the app crashes when the user presses anywhere outside the actionsheet. The reason for this is that the value of `selectedIndex` when a user tries to cancel on android is -1, and the value of `selectedIndex` when a user tries to cancel on iOS is 0. This is due to

Test Plan: No longer triggering the crash, and we are getting the expected cancel behavoir

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7456