[native] add safe area insets for android action sheet
Summary: Another thing I encountered when testing the action sheet for the edit avatar flow on android is that if the nav buttons are visible then it partly covers the most bottom option on the action sheet. To fix this I used `SafeAreaContext` to set the `paddingBottom` of the action sheet which fixes the issue.

Test Plan: Please check out the screenshots to see the changes I made

Reviewers:

Subscribers: