[native] fix Android cancel action sheet crash
Summary: While testing the edit avatar flow for android, I encountered this bug where the app crashes when the user presses anywhere outside the actionsheet. The reason for this is that the value of `selectedIndex` when a user tries to cancel on android is -1, and the value of `selectedIndex` when a user tries to cancel on iOS is 0. This is due to action sheets on android [[ https://m2.material.io/components/sheets-bottom#usage | generally not having a cancel option in the sheet ]] but [[ https://developer.apple.com/design/human-interface-guidelines/components/presentation/action-sheets/ | there being a cancel option with iOS ]]

Test Plan:
No longer triggering the crash, and we are getting the expected cancel behavior. Also double checked that iOS still behaves as expected

Before:

{F488563}

After:
{F488568}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7456