[native] add safe area insets for android action sheet
Summary:
Another thing I encountered when testing the action sheet for the edit avatar flow on android is that if the nav buttons are visible then it partly covers the most bottom option on the action sheet. To fix this I used `SafeAreaContext` to set the `paddingBottom` of the action sheet which fixes the issue.

Depends on D7456

Test Plan:
Please check out the screenshots to see the changes I made

Before:
{F488596}

After:
{F488597}

w/o nav buttons:
{F488598}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7457