[services][blob] Introduce HTTP context utils
Summary:

This diff introduces a `AppContext` structure that is passed to each HTTP request handler. It contains references
to database and S3 instances along with some utilities that will be used in the handlers.

Depends on D7444

Test Plan:

This is not yet used, it will be tested with the next diff.

Reviewers: marcin, tomek, jon, michal

Subscribers: