[services][blob] Implement Delete blob HTTP handler
Summary:

Resolves [[ https://linear.app/comm/issue/ENG-3524/blob-service-implement-remove-blob-http-endpoint | ENG-3524 ]]

THis diff implements the `DELETE /blob/{holder}` endpoint which revokes a blob holder. If last holder is revoked, the blob is deleted. This behavior is the same as for `Remove` blob RPC.

Depends on D7465

Test Plan:

Created a blob with a few holders and called this endpoint to remove them one by one.
Calling `GET` before returns HTTP 200, after deleting returns HTTP 404 as expected.

Reviewers: marcin, tomek, jon, michal

Subscribers: