[keyserver] Introduce validateOutput function
Summary:
Introduces `validateOutput` function that validates and (right now blocked with a flag) converts the ids of the responses. Also introduces a pair of functions `convertServerIDsToClientIDs` (used in `validateOutput` and in tests) and `convertClientIDsToServerIDs` (only used in tests right now, will be used in the future diff for the input conversion).

Depends on D7710

Test Plan:
- `yarn jest`
- Test if the app works fine after adding `validateOutput` to responders in the later diffs

Reviewers: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7711