[native] Left-align pending thread header titles
Summary: We currently center-align a `MessageListHeaderTitle` if it is `searchEmpty` ("New Message" with no users selected), or if it is a pending thread (see definition of `areSettingsEnabled`).

Center-aligned the header title for pending threads is inconsistent with behavior for normal threads. Discussed this with Ted and @ginsu and we decided it would be more consistent to always left-align, except in the "New Message" case.

Test Plan: [videos]

Reviewers: ginsu, atul

Subscribers: