[native] Left-align pending thread header titles
Summary:
We currently center-align a `MessageListHeaderTitle` if it is `searchEmpty` ("New Message" with no users selected), or if it is a pending thread (see definition of `areSettingsEnabled`).

Center-aligned the header title for pending threads is inconsistent with behavior for normal threads. Discussed this with Ted and @ginsu and we decided it would be more consistent to always left-align, except in the "New Message" case.

Some previous context in D7134. Linear task: [ENG-3693](https://linear.app/comm/issue/ENG-3693/center-align-messagelistheadertitle-for-pending-threads)

Test Plan:
| before | after |
| {F490838} | {F490839} |

Reviewers: ginsu, atul

Reviewed By: ginsu

Subscribers: tomek, ginsu

Differential Revision: https://phab.comm.dev/D7474