[native] Process and upload image for image avatar in `EditAvatar`
Summary:
We're able to process and upload image for image avatar and get back `uploadID`. However, before we're able to include `uploadID` in an image avatar update request we'll need additional information about thread/user passed into `EditAvatar`. It looks like the actual Avatars are being displayed as child of `EditAvatar`... I wonder if it makes sense to move `UserAvatar` and `ThreadAvatar` "within` `EditAvatar` and display them conditionally based on `props`?

Test Plan:
Logged result of `processMedia` and observed that it was as expected.

Checked `uploads` table and saw uploaded images:

Reviewers:
ashoat, ginsu