[native] Add horizontal margin to buttons in LoggedOutModal
Summary:
In order to iterate on the new registration flow, I'm going to replace the register button with two side-by-side register buttons, with the second one only visible to devs.

In order to make it easier to add space between those side-by-side buttons, I'm adding some horizontal margin to the buttons. (This [wouldn't be necessary](https://reactnative.dev/blog/2023/01/12/version-071#simplifying-layouts-with-flexbox-gap) on React Native 0.71, where we could use `gap` properties instead of `margin` ones.)

Test Plan:
[screenshots]

The only difference is in the horizontal margins of `LoggedOutStaffInfo`

Reviewers: ginsu, atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7494