[native] Move complexity of `actionSheetConfig` to `useShowAvatarActionSheet`
Summary:
We can simplify the `actionSheetConfig` constructed in `Edit*Avatar` by moving some of what's common to `useShowAvatarActionSheet`. This cuts down on reptition a decent amount.

All of this refactoring brought the size of `Edit*Avatar` from  234 lines to 40:

This should make the subsequent changes easier to manage.

---

Depends on D7506

Test Plan:
Everything looks and works as it did previously + ESLint + Flow

Reviewers:
ashoat, ginsu