[native] Show `ActivityIndicator` in `EditUserAvatar` during update
Summary:
We consume `isGalleryAvatarUpdateLoading` and `isRemoveAvatarUpdateLoading` introduced in previous diffs to display `ActivityIndicator` (ie spinner) when an update is in progress.

Note: Yeah there are obvious things that can be factored out. I want to get more of the error handling stuff done first and then I'll go through and consolidate things.

---

Depends on D7528

Test Plan: Works as expected:

Reviewers: ashoat, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7529