[native] Replace logging with `Alert`s in `selectFromGalleryAndUpdateUserAvatar`
Summary:
Replace `console.log` with `Alert`s so if something goes wrong with media selection/processing/uploading we can at least give the user an indication that the operation failed.

Proper error handling would include cleaning up temporary files + creating a media mission report, but deferring that for now.

NOTE: Currently if `dispatchActionPromise` of `updateUserAvatarCall` fails it'll fail silently. One possibility could be to wrap the call like we do in `siwe-login-form` to introspect (see below), or we could have a `useEffect` that `Alert.alert(...)`s if `loadingStatus === 'error'`?

{F496369}

---

Depends on D7529

Test Plan:
Flipped boolean and observed that Alert appeared as expected:

{F496367}

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7530