[native] Add a pin/unpin icon in the tooltip
Summary:
The tooltip icon to pin / unpin messages should appear for
composable messages (multimedia and text, not robotext). Functionality for
pressing the icon will be added in the next diff.

Note: These icons do need to be modified a bit, I believe Ted said that
the icons for native should just be the outline and unfilled.

Test Plan:
Confirm that the tooltip icons appear for the messages that
should have them

Reviewers: atul, ginsu, tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7606