[native] Handle error state when joining a community
Summary:
When an error occurs, we're updating the params to let the user know that the link is invalid. This isn't ideal, as the error might be caused by e.g. a timeout and a link is still valid, but it is by far the simplest solution.

Depends on D7558

Test Plan: Introduce a sleep on server side and the throw. Check if the error state becomes enabled.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7559