[web] Show message tooltip on hover in the pinned messages modal
Summary:
Currently, the tooltip displays below the pinned messages modal when a message is hovered over. This is because of the z-index property in the [[
https://github.com/CommE2E/comm/blob/master/lib/components/modal-overlay.css#L6 | ModalOverlay]] component.

The solution here is to make the tooltip container, icons, and label to have a z-index of 5 so it displays above the modal. This visually has no effect on the tooltip in chat.

Linear: https://linear.app/comm/issue/ENG-3451/show-message-tooltip-on-hover-in-the-pinned-messages-modal

Depends on D7505

Test Plan: Please watch the video showing the behavior of the tooltip in both chat and the pinned messages modal

Reviewers: tomek, ginsu, atul

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D7563