[native] Create the toggle pin modal
Summary: When pressing the icon in the tooltip, a modal should pop up on the screen allowing the user to pin / unpin a message. This diff handles the modal creation, but to make it easier to review, the
actual rendering of the message in the modal will be in the next diff following this one.

I placed the modal in `app-navigator.react.js` since the `TextMessage` component requires the `OverlayContext` to be set, so when rendering the `Message > TextMessage` inside of this modal in the next diff,
we won't have the `invariant` requiring this context be thrown.

Note: The approach here on `native` follows closely the approach for `web` here: https://phab.comm.dev/D7309

Depends on D7606

Test Plan: Confirm that the modal appears (not rendering the message yet) for messages and multimedia, both that should be pinned and unpinned.

Reviewers: atul, ginsu, tomek

Subscribers: