[lib] Move `getRolePermissionBlobs` to `lib`
Summary:
We're going to use `getRolePermissionBlobs` to calculate thread permissions on the client "from first principles" (ie based on threadTypes and `getRolePermissionBlobs` without needing roles/permissions from the `keyserver`).

This diff just moves `getRolePermissionBlobs` to `lib` so we can use it from `native`.

Test Plan:
This should be a super straightforward drag and drop diff.

However, since it's consumed by `updateRolesAndPermissionsForAllThreads` I kicked off a `updateRolesAndPermissionsForAllThreads` migration on `keyserver` and made sure that roles/permissions continued to be calculated as expected... though it would be very surprising if it didn't.

Reviewers: ginsu, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7581