[native] fix rendering of edit thread name settings button
Summary:
As I was testing the loaders in Thread Settings for [[ https://linear.app/comm/issue/ENG-3741/change-thread-settings-loading-state-are-shared-across-different | ENG-3741 ]] I came across an issue where if the user navigated away from thread settings and back to it, then the button would not render while the new name was being saved. This was happening because `nameEditValue` was not being set since we navigated away from the screen so we were only just rendering `SaveSettingButton` which in turn render nothing since `canChangeSettings` is set to false while the name is being set. To fix this I created a separate function that just handles the rendering of the button and used that function within `renderContent` and now an `ActivityIndicator` will always render while the setting is being saved

A follow up diff for fixing the description will be up shortly

Test Plan:
Please see the demo videos below to see the changes I made:

Before:
{F505409}

After:
{F505411}

Reviewers: ashoat, atul

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7588